requirements.txt
Flask
PyMuPDF
google-generativeai
uuid
python-dotenv
